+++ orgbox.el
@@ -1,11 +1,11 @@
-;;; orgbox.el --- Mailbox-like task scheduling in org agenda
+;;; orgbox.el --- Mailbox-like task scheduling in org agenda.
 
 ;; Copyright © 2014 Yasuhito Takamiya <yasuhito@gmail.com>
 
 ;; Author: Yasuhito Takamiya <yasuhito@gmail.com>
 ;; URL: https://github.com/yasuhito/orgbox
 ;; Keywords: org
-;; Version: 0.2.0
+;; Version: 0.3.0
 ;; Package-Requires: ((org "8.0") (cl-lib "0.5"))
 
 ;; This file is not part of Org.
@@ -46,29 +46,122 @@
 (require 'cl-lib)
 (require 'org-agenda)
 
-(defun orgbox-later-today ()
+(defvar orgbox-start-time-of-day)
+(defvar orgbox-start-day-of-week)
+(defvar orgbox-start-time-of-weekends)
+(defvar orgbox-start-day-of-weekends)
+(defvar orgbox-start-time-of-evening)
+(defvar orgbox-later)
+(defvar orgbox-someday)
+
+(defgroup orgbox nil
+  "Mailbox-like task scheduling in org agenda."
+  :group 'org)
+
+(defcustom orgbox-start-time-of-day "8:00"
+  "What time does your day start?"
+  :group 'orgbox
+  :type 'string)
+
+(defcustom orgbox-start-day-of-week "mon"
+  "What day does your week start?"
+  :group 'orgbox
+  :type '(choice (const :tag "Monday" "mon")
+                 (const :tag "Tuesday" "tue")
+                 (const :tag "Wednesday" "wed")
+                 (const :tag "Thursday" "thu")
+                 (const :tag "Friday" "fri")
+                 (const :tag "Saturday" "sat")
+                 (const :tag "Sunday" "sun")))
+
+(defcustom orgbox-start-time-of-weekends "10:00"
+  "What time does your weekends start?"
+  :group 'orgbox
+  :type 'string)
+
+(defcustom orgbox-start-day-of-weekends "sat"
+  "What day does your weekends start?"
+  :group 'orgbox
+  :type '(choice (const :tag "Monday" "mon")
+                 (const :tag "Tuesday" "tue")
+                 (const :tag "Wednesday" "wed")
+                 (const :tag "Thursday" "thu")
+                 (const :tag "Friday" "fri")
+                 (const :tag "Saturday" "sat")
+                 (const :tag "Sunday" "sun")))
+
+(defcustom orgbox-start-time-of-evening "18:00"
+  "What time does your evening start?"
+  :group 'orgbox
+  :type 'string)
+
+(defcustom orgbox-later 3
+  "Specify 'later' in number of hours."
+  :type '(choice (const :tag "1 Hour" 1)
+                 (const :tag "2 Hours" 2)
+                 (const :tag "3 Hours" 3)
+                 (const :tag "4 Hours" 4)
+                 (const :tag "5 Hours" 5)
+                 (const :tag "6 Hours" 6)
+                 (const :tag "7 Hours" 7)
+                 (const :tag "8 Hours" 8)
+                 (const :tag "9 Hours" 9)
+                 (const :tag "10 Hours" 10)
+                 (const :tag "11 Hours" 11)
+                 (const :tag "12 Hours" 12)
+                 (const :tag "13 Hours" 13)
+                 (const :tag "14 Hours" 14)
+                 (const :tag "15 Hours" 15)
+                 (const :tag "16 Hours" 16)
+                 (const :tag "17 Hours" 17)
+                 (const :tag "18 Hours" 18)
+                 (const :tag "19 Hours" 19)
+                 (const :tag "20 Hours" 20)
+                 (const :tag "21 Hours" 21)
+                 (const :tag "22 Hours" 22)
+                 (const :tag "23 Hours" 23)
+                 (const :tag "24 Hours" 24))
+  :group 'orgbox)
+
+(defcustom orgbox-someday "+3m"
+  "Specify 'Someday' in number of months."
+  :type '(choice (const :tag "1 Month" "+1m")
+                 (const :tag "2 Months" "+2m")
+                 (const :tag "3 Months" "+3m")
+                 (const :tag "4 Months" "+4m")
+                 (const :tag "5 Months" "+5m")
+                 (const :tag "6 Months" "+6m")
+                 (const :tag "7 Months" "+7m")
+                 (const :tag "8 Months" "+8m")
+                 (const :tag "9 Months" "+9m")
+                 (const :tag "10 Months" "+10m")
+                 (const :tag "11 Months" "+11m")
+                 (const :tag "12 Months" "+12m"))
+  :group 'orgbox)
+
+(defun orgbox-schedule-later-today ()
   "Schedule a task for later today."
   (interactive)
-  (let ((later-today (format-time-string "%Y-%m-%d %H:%M"
-                                           (time-add (current-time)
-                                                     (seconds-to-time (* 3 60 60))))))
-    (org-agenda-schedule nil later-today)))
+  (let ((later (format-time-string "%Y-%m-%d %H:%M"
+                                   (time-add (current-time)
+                                             (seconds-to-time (* orgbox-later 60 60))))))
+    (org-agenda-schedule nil later)))
 
 (defun orgbox-evening-p ()
   "Is already evening?"
-  (> (nth 2 (decode-time (current-time))) 18))
+  (string< orgbox-start-time-of-evening (format-time-string "%H:%M" (current-time))))
 
-(defun orgbox-this-or-tomorrow-evening ()
+(defun orgbox-schedule-this-or-tomorrow-evening ()
   "Schedule a task for this or tomorrow evening."
   (interactive)
   (if (orgbox-evening-p)
-      (org-agenda-schedule nil "+1d 18:00")
-    (org-agenda-schedule nil "18:00")))
+      (org-agenda-schedule nil (format "+1d %s" orgbox-start-time-of-evening))
+    (org-agenda-schedule nil orgbox-start-time-of-evening)))
 
-(defun orgbox-tomorrow ()
+(defun orgbox-schedule-tomorrow ()
   "Schedule a task for tomorrow."
   (interactive)
-  (org-agenda-schedule nil "+1d 8:00"))
+  (org-agenda-schedule nil (format "+1d %s" orgbox-start-time-of-day)))
 
 (defun orgbox-weekend-p ()
   "Today is weekend?"
@@ -76,25 +169,25 @@
                       (calendar-gregorian-from-absolute (org-today)))))
     (member day-of-week org-agenda-weekend-days)))
 
-(defun orgbox-this-or-next-weekend ()
+(defun orgbox-schedule-this-or-next-weekend ()
   "Schedule a task for this or next weekend."
   (interactive)
-  (org-agenda-schedule nil "+sat 10:00"))
+  (org-agenda-schedule nil (format "+%s %s" orgbox-start-day-of-weekends orgbox-start-time-of-weekends)))
 
-(defun orgbox-next-week ()
+(defun orgbox-schedule-next-week ()
   "Schedule a task for next week."
   (interactive)
-  (org-agenda-schedule nil "+mon 8:00"))
+  (org-agenda-schedule nil (format "+%s %s" orgbox-start-day-of-week orgbox-start-time-of-day)))
 
-(defun orgbox-in-a-month ()
+(defun orgbox-schedule-in-a-month ()
   "Schedule a task for 1 month later."
   (interactive)
   (org-agenda-schedule nil "+1m"))
 
-(defun orgbox-someday ()
+(defun orgbox-schedule-someday ()
   "Schedule a task for someday."
   (interactive)
-  (org-agenda-schedule nil "+3m"))
+  (org-agenda-schedule nil orgbox-someday))
 
 (defun orgbox ()
   "Schedule a task interactively."
@@ -105,16 +198,16 @@
            (if (orgbox-weekend-p) "next" "this"))
   (let ((a (read-char-exclusive)))
     (cl-case a
-      (?l (call-interactively 'orgbox-later-today))
-      (?e (call-interactively 'orgbox-this-or-tomorrow-evening))
-      (?t (call-interactively 'orgbox-tomorrow))
-      (?w (call-interactively 'orgbox-this-or-next-weekend))
-      (?n (call-interactively 'orgbox-next-week))
-      (?m (call-interactively 'orgbox-in-a-month))
-      (?s (call-interactively 'orgbox-someday))
+      (?l (call-interactively 'orgbox-schedule-later-today))
+      (?e (call-interactively 'orgbox-schedule-this-or-tomorrow-evening))
+      (?t (call-interactively 'orgbox-schedule-tomorrow))
+      (?w (call-interactively 'orgbox-schedule-this-or-next-weekend))
+      (?n (call-interactively 'orgbox-schedule-next-week))
+      (?m (call-interactively 'orgbox-schedule-in-a-month))
+      (?s (call-interactively 'orgbox-schedule-someday))
       (?p (call-interactively 'org-agenda-schedule))
       (?q (message "Abort"))
-      (otherwise (error "Invalid key" )))))
+      (otherwise (error "Invalid key")))))
 
 (org-defkey org-agenda-mode-map (kbd "C-c C-s") 'orgbox)